Patch for 64 bit

+++ src/NotificationAreaIcon.cs
@@ -403,7 +403,7 @@ namespace Banshee
 	internal struct XClientMessageEvent
 	{
 		internal XEventName     type;
-		internal int            serial;
+		internal IntPtr         serial;
 		internal bool           send_event;
 		internal IntPtr         display;
 		internal IntPtr         window;
